+++ docs/conf.py
@@ -1,64 +1,39 @@
-# -- General configuration -----------------------------------------------------
 import os
 import sys
+
+try:
+    import furo
+except ImportError:
+    furo = None
+try:
+    import jaraco.packaging.sphinx as jaraco_packaging_sphinx
+except ImportError:
+    jaraco_packaging_sphinx = None
+try:
+    import rst.linker as rst_linker
+except ImportError:
+    rst_linker = None
 
 # Use the source tree.
 sys.path.insert(1, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
 
-# Add any Sphinx extension module names here, as strings. They can be extensions
-# coming with Sphinx (named 'sphinx.ext.*') or your custom ones.
 extensions = [
     'sphinx.ext.autodoc',
     'sphinx.ext.doctest',
-    'sphinx.ext.intersphinx',
     'sphinx.ext.todo',
     'sphinx.ext.coverage',
-    'jaraco.packaging.sphinx',
-    'rst.linker',
 ]
+if jaraco_packaging_sphinx is not None:
+    extensions += ['jaraco.packaging.sphinx']
 
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-
-# The suffix of source filenames.
-source_suffix = '.rst'
-
-# The master toctree document.
 master_doc = 'index'
 
-# General information about the project.
-copyright = u'2008-2011, John Paulett; 2009-2020, David Aguilar'
+if furo is not None:
+    html_theme = 'furo'
 
-# List of directories, relative to source directory, that shouldn't be searched
-# for source files.
-exclude_trees = []
-
-# The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
-
-# -- Options for HTML output ---------------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  Major themes that come with
-# Sphinx are currently 'default' and 'sphinxdoc'.
-html_theme = 'default'
-
-# Output file base name for HTML help builder.
-htmlhelp_basename = 'jsonpickledoc'
-
-
-# -- Options for LaTeX output --------------------------------------------------
-
-# Grouping the document tree into LaTeX files. List of tuples
-# (source start file, target name, title, author, documentclass [howto/manual]).
-latex_documents = [
-    ('index', 'jsonpickle.tex', 'jsonpickle Documentation', 'David Aguilar', 'manual'),
-]
-
-
-# Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/3': None}
-
-
+# Link dates and other references in the changelog
+if rst_linker is not None:
+    extensions += ['rst.linker']
 link_files = {
     '../CHANGES.rst': dict(
         using=dict(GH='https://github.com'),
@@ -72,7 +47,7 @@ link_files = {
                 url='{package_url}/pull/{pull}',
             ),
             dict(
-                pattern=r'^(?m)((?P<scm_version>v?\d+(\.\d+){1,2}))\n[-=]+\n',
+                pattern=r'(?m:^((?P<scm_version>v?\d+(\.\d+){1,2}))\n[-=]+\n)',
                 with_scm='{text}\n{rev[timestamp]:%d %b %Y}\n',
             ),
             dict(
@@ -82,3 +57,17 @@ link_files = {
         ],
     )
 }
+
+# Be strict about any broken references
+nitpicky = True
+
+sphinx_disable = os.environ.get('JSONPICKLE_SPHINX_DISABLE', '')
+if 'intersphinx' not in sphinx_disable:
+    extensions += ['sphinx.ext.intersphinx']
+intersphinx_mapping = {
+    'python': ('https://docs.python.org/3', None),
+    'sphinx': ('https://www.sphinx-doc.org/en/stable/', None),
+}
+
+# Preserve authored syntax for defaults
+autodoc_preserve_defaults = True