+++ jsonpickle/tags.py
@@ -6,6 +6,7 @@ created by the Pickler class.  The Unpickler uses
 these custom key names to identify dictionaries
 that need to be specially handled.
 """
+
 from __future__ import absolute_import, division, unicode_literals
 
 BYTES = 'py/bytes'